requirements.txt
boto3==1.28.44
chromadb==0.4.9
langchain==0.0.292
python-dotenv==1.0.0
PyYAML==6.0.1
streamlit==1.26.0
sentence-transformers==2.2.2
openai==0.28.0
unstructured==0.10.15
markdown==3.4.4
aws-lambda-powertools==2.25.0
psycopg2-binary==2.9.7
pgvector==0.2.2
pypdf==3.16.4